# improve(planner): optimize exists subquery