I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

1. optimize exists subquery
2. fix `limit` display

Before:
```sql
mysql> select  avg(a.number) from numbers(1000000) as a  where EXISTS ( select number from numbers(1000000000));
+---------------+
| avg(a.number) |
+---------------+
|      499999.5 |
+---------------+
1 row in set (5.48 sec)
Read 1001000000 rows, 7.46 GiB in 5.442 sec., 183.96 million rows/sec., 1.37 GiB/sec.
```

Now:
```sql
mysql> select  avg(a.number) from numbers(1000000) as a  where EXISTS ( select number from numbers(1000000000));
+---------------+
| avg(a.number) |
+---------------+
|      499999.5 |
+---------------+
1 row in set (3.13 sec)
Read 1100000 rows, 8.39 MiB in 3.099 sec., 355.01 thousand rows/sec., 2.71 MiB/sec.
```

## Changelog

- Improvement

## Related Issues

Fixes #issue