@sundy-li 2022-06-30
/LGTM 

But the best result of the optimization is remove the exist subquery because it's always true and there is no correlated subquery.