@xudong963 2022-06-30
> /LGTM
> 
> But the best result of the optimization is to remove the exist subquery because it's always true and there is no correlated subquery.

Yes, will consider to some add fast paths for some specific cases later. The PR is an optimization for any exists subquery.